requirements.txt
alembic==1.9.4
asttokens==2.2.1
backcall==0.2.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
executing==1.2.0
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
ipdb==0.13.11
ipython==8.10.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
matplotlib-inline==0.1.6
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.36
pure-eval==0.2.2
Pygments==2.14.0
six==1.16.0
SQLAlchemy==2.0.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.5.0
wcwidth==0.2.6
Werkzeug==2.2.3
pytest-mock
dynaconf==3.1.7
